+++ Rasa_Bot/actions/actions_common.py
@@ -1,10 +1,10 @@
 import logging
 
 from celery import Celery
+from niceday_client import NicedayClient, definitions
 from rasa_sdk import Action
 from rasa_sdk.events import FollowupAction
-from .definitions import REDIS_URL
-
+from .definitions import REDIS_URL, NICEDAY_API_ENDPOINT
 
 celery = Celery(broker=REDIS_URL)
 
@@ -37,3 +37,34 @@ class MarkDialogAsCompleted(Action):
         logging.info("no celery error")
 
         return []
+
+
+class GetMetadata(Action):
+    def name(self):
+        return "action_build_metadata"
+
+    async def run(self, dispatcher, tracker, domain):
+        dispatcher.utter_message(
+            json_message={"text": "message",
+                          "attachmentIds": [1733]},
+        )
+        return[]
+
+
+class UploadFile(Action):
+    def name(self):
+        return "action_upload_file"
+
+    async def run(self, dispatcher, tracker, domain):
+        import os
+        cwd = os.getcwd()
+        logging.info("CURRENT DIRECTORY")
+        logging.info(cwd)
+        client = NicedayClient(NICEDAY_API_ENDPOINT)
+        user_id = int(tracker.current_state()['sender_id'])
+
+        file = open('/app/tst.PNG', 'rb')
+        filepath = '/app/tst.PNG'
+
+        response = client.upload_file(user_id, 'tst.png', filepath, file)
+        return[]
